refactor(headless-crawler): jobDetail関連ロジックのディレクトリ統合・整理
- jobDetailPage配下のextractor/loader/transformer/helpersをjobDetail/配下に統合
- 旧E-T-scraper系・config系のファイルを削除
- handler等のimportパスを新構成に合わせて修正
- コード構成の可読性・保守性向上
