+++ apps/headless-crawler/lib/jobDetail/extractor/index.ts
