+++ apps/headless-crawler/lib/jobDetail/helpers/checkers/error.ts
