+++ apps/headless-crawler/lib/jobDetail/helpers/checkers/index.ts
