+++ apps/headless-crawler/lib/jobDetail/helpers/extractors/error.ts
