+++ apps/headless-crawler/lib/jobDetail/helpers/extractors/index.ts
