+++ apps/headless-crawler/lib/jobDetail/helpers/validators/error.ts
