+++ apps/headless-crawler/lib/jobDetail/helpers/validators/index.ts
