+++ apps/headless-crawler/lib/jobDetail/loader/helper.ts
