+++ apps/headless-crawler/lib/jobDetail/transformer/helper.ts
